+++ src/layout/Header.js
@@ -1,9 +1,9 @@
 import Link from "next/link";
 import { Fragment, useEffect, useState } from "react";
 import { activeSection } from "../utilits";
-import { Main } from "next/document";
-// import Image from "next/image";
+
 const Header = ({ blog, projects }) => {
+  
   const [sideBarToggle, setSideBarToggle] = useState(false);
   useEffect(() => {
     if (!blog && !projects) {
